Build and run plugin via containers.
Because the official contianers are only availalbe as AMD64 on the registry, I have built my own container from their sources which runs on ARM64 (my M1 MacBook).

This will make things more complicated later, when we're building for multiple architectures. Perhaps that should just happen on CI as part of the release?