Hacked together a working test.
We use the Application Token in the Webhook URL like Slack does it. Since the Slack Webhook URLs contian secrets (and are therefor considered secrest), I expect applications to take precautions around this value already.

There is no simple API to send messages on behalf of another App. So we simply sent the notifiaction against the local REST API with Markdown formating. There are configuration options where this will fail (namely when port/endpoint are specified outside of ENV variables). These should be in the readme AND the message.