Unify: XElement blockkits don't write their own newlines.
This is for simplicity and consistency across the implementations. In general, Markdown block elements should take care to render themselves as valid, which will require them to also render newlines. The Elements are however inline, so they do not require their own newline to be valid.

This allows them to be used inline by other blocks.