Extracted Gotify client to send messages on behalv of apps.
The standard Plugin SDK only lets you send messages on behalf of the Plugin App. We want to instead allow sending messages as _any_ existing app by adding the apps token into the webhook URL. This requires us to make a web request to the local Gotify servers REST API, which this client does for us.