FIX: Correct release upload file pattern.
The files cretaed by "Upload Artifact" are actually ZIP files, not the raw SO files. Changed their names accordingly. Also used the wildcard at the end of the artifact name rather than the beginning, because I'm not sure if these don't contain a hidden `.zip` at the end - that could be why they don't match.