CI build/publish a Docker Image with plugin and server bundled.
This makes using this plugin on Kubernetes (for my case) much easier, as it only needs to refer to this image.

Also, with the help of https://github.com/docker/buildx/discussions/1928 this is now a true multi-arch image, no need to target a specific image-name for your platform (the correct platform image will automatically be downloaded).