Generate an attestation for Docker image.
This generates proof that the conainter image was actually created using the CI run. Its supposed to increase trust in image consumers and allow them to verify there was no hidden supply-chain attack, since they can always look at the CI run that created the image and at the Repo at the time of image creation.